Fix catch rates on Lure, Safari and Sport balls
None of these were working as they should due to a misunderstanding of changes made in Gen 8 & 9. To uphold parity with their use as cosmetic balls, Safari and Sport balls retain their old 1.5x rate when used in their respective battle types. Lure balls were only not working as intended in SV due to an oversight (surprise surprise)-- the changes made upon reimplementation are due to the lack of fishing as a mechanic. There could be further discussion on having a config to give them their modern functionality (being identical to Dive Balls), but this is the best for a quick fix.