Modernize every effect script
Modernized Fling behaviour to allow the Griseous Orb to be flung and for the move to work when used by Pokemon with Multitype (Arceus).

Fixed Judgment and Multi-Attack calling TYPE_MYSTERY for Pixie Plate instead of TYPE_FAIRY (both have an ID of 9, though in some cases the latter is different).

Fixed Acrobatics calling BSCRIPT_VAR_BATTLE_TYPE instead of ITEM_NONE (both have an ID of 0)
Fixed typo in the name of effect script 318 (Geomancy) that indicated it raised Atk instead of SpDef.

Fixed oversight in effect script 321 (Infernal Parade) where Comatose was not being checked.